+++ src/registry/mod.rs
@@ -6,23 +6,37 @@ pub mod server;
 mod storage;
 mod temporary_file;
 
-use std::path::PathBuf;
-
 pub use admission::ImageValidationConfig;
-use anyhow::Result;
 pub use blob_storage::{BlobReader, ContentInfo, UploadInfo};
 pub use digest::Digest;
 pub use manifest_storage::ManifestReader;
-pub use proxy::{RegistryProxiesConfig, SingleRegistryProxyConfig};
+pub use proxy::{RegistryProxiesConfig, RemoteImage, SingleRegistryProxyConfig};
+use serde::Deserializer;
 pub use server::TrowServer;
 pub use storage::StorageBackendError;
 use thiserror::Error;
-use tracing::{event, Level};
 
 pub mod blob_storage;
 #[allow(dead_code)]
 pub mod digest;
 pub mod manifest_storage;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, Debug, Default)]
+pub struct ConfigFile {
+    #[serde(deserialize_with = "de_unwrap_or_default")]
+    pub registry_proxies: RegistryProxiesConfig,
+    pub image_validation: Option<ImageValidationConfig>,
+}
+
+fn de_unwrap_or_default<'de, T, D>(d: D) -> Result<T, D::Error>
+where
+    D: Deserializer<'de>,
+    T: Deserialize<'de> + Default,
+{
+    Deserialize::deserialize(d).map(|x: Option<_>| x.unwrap_or_default())
+}
 
 // Storage Driver Error
 #[derive(Error, Debug)]
@@ -45,7 +59,7 @@ pub enum RegistryError {
 
 impl From<sqlx::Error> for RegistryError {
     fn from(err: sqlx::Error) -> Self {
-        event!(Level::ERROR, "Database error: {err:?}");
+        tracing::error!("Database error: {err:?}");
         Self::Internal
     }
 }
@@ -55,46 +69,17 @@ impl From<StorageBackendError> for RegistryError {
         match err {
             StorageBackendError::BlobNotFound(_) => Self::NotFound,
             StorageBackendError::Internal(e) => {
-                event!(Level::ERROR, "Internal storage error: {e}");
+                tracing::error!("Internal storage error: {e}");
                 Self::Internal
             }
             StorageBackendError::InvalidContentRange => Self::InvalidContentRange,
             StorageBackendError::InvalidDigest => Self::InvalidDigest,
-            StorageBackendError::InvalidManifest(_msg) => Self::InvalidManifest,
             StorageBackendError::InvalidName(name) => Self::InvalidName(name),
             StorageBackendError::Io(e) => {
-                event!(Level::ERROR, "Internal IO error: {e:?}");
+                tracing::error!("Internal IO error: {e:?}");
                 Self::Internal
             }
             StorageBackendError::Unsupported => Self::Unsupported,
         }
-    }
-}
-
-pub struct TrowServerBuilder {
-    data_path: PathBuf,
-    proxy_registry_config: Option<RegistryProxiesConfig>,
-    image_validation_config: Option<ImageValidationConfig>,
-}
-
-pub fn build_server(
-    data_path: PathBuf,
-    proxy_registry_config: Option<RegistryProxiesConfig>,
-    image_validation_config: Option<ImageValidationConfig>,
-) -> TrowServerBuilder {
-    TrowServerBuilder {
-        data_path,
-        proxy_registry_config,
-        image_validation_config,
-    }
-}
-
-impl TrowServerBuilder {
-    pub async fn get_server(self) -> Result<TrowServer> {
-        TrowServer::new(
-            self.data_path,
-            self.proxy_registry_config,
-            self.image_validation_config,
-        )
     }
 }