+++ src/registry/proxy/proxy_config.rs
@@ -1,17 +1,14 @@
-use anyhow::Result;
 use aws_config::BehaviorVersion;
 use aws_sdk_ecr::config::http::HttpResponse;
 use aws_sdk_ecr::error::SdkError;
 use aws_sdk_ecr::operation::get_authorization_token::GetAuthorizationTokenError;
 use base64::Engine;
-use bytes::Bytes;
 use futures::future::try_join_all;
 use oci_client::client::ClientProtocol;
 use oci_client::secrets::RegistryAuth;
 use oci_client::Reference;
 use serde::{Deserialize, Serialize};
 use sqlx::SqlitePool;
-use tracing::{event, Level};
 
 use crate::registry::manifest::{ManifestReference, OCIManifest};
 use crate::registry::proxy::remote_image::RemoteImage;
@@ -80,10 +77,23 @@ pub enum DownloadRemoteImageError {
     InvalidDigest(#[from] DigestError),
     #[error("Failed to download image")]
     DownloadAttemptsFailed,
+    #[error("Manifest JSON is not canonicalized")]
+    ManifestNotCanonicalized,
+    #[error("OCI client error: {0}")]
+    OciClientError(#[from] oci_client::errors::OciDistributionError),
+    #[error("Storage backend error: {0}")]
+    StorageError(#[from] crate::registry::storage::StorageBackendError),
+    #[error("Could not deserialize manifest: {0}")]
+    ManifestDeserializationError(#[from] serde_json::Error),
+    #[error("Could not get AWS ECR password: {0}")]
+    EcrLoginError(#[from] EcrPasswordError),
 }
 
 impl SingleRegistryProxyConfig {
-    async fn setup_client(&self, insecure: bool) -> Result<(oci_client::Client, RegistryAuth)> {
+    async fn setup_client(
+        &self,
+        insecure: bool,
+    ) -> Result<(oci_client::Client, RegistryAuth), DownloadRemoteImageError> {
         let mut client_config = oci_client::client::ClientConfig::default();
         if insecure {
             client_config.protocol = ClientProtocol::Http;
@@ -109,10 +119,10 @@ impl SingleRegistryProxyConfig {
         image: &RemoteImage,
         registry: &TrowServer,
         db: &SqlitePool,
-    ) -> Result<Digest, DownloadRemoteImageError> {
+    ) -> Result<String, DownloadRemoteImageError> {
         // Replace eg f/docker/alpine by f/docker/library/alpine
         let repo_name = format!("f/{}/{}", self.alias, image.get_repo());
-        event!(Level::DEBUG, "Downloading proxied image {}", repo_name);
+        tracing::debug!("Downloading proxied image {}", repo_name);
 
         let image_ref: Reference = image.clone().into();
         let try_cl = self.setup_client(image.scheme == "http").await.ok();
@@ -141,7 +151,7 @@ impl SingleRegistryProxyConfig {
                                 digests.push(Digest::try_from_raw(&d)?);
                             }
                         }
-                        Err(e) => event!(Level::WARN, "Failed to fetch manifest digest: {}", e),
+                        Err(e) => tracing::warn!("Failed to fetch manifest digest: {}", e),
                     }
                 }
                 if let Some(local_digest) = local_digest {
@@ -153,20 +163,14 @@ impl SingleRegistryProxyConfig {
 
         for mani_digest in digests.into_iter() {
             let mani_digest_str = mani_digest.as_str();
-            let have_manifest = sqlx::query_scalar!(
-                r#"
-                SELECT EXISTS(
-                    SELECT 1 FROM blob
-                    WHERE digest = $1 AND is_manifest = 1
-                )
-                "#,
+            let has_manifest = sqlx::query_scalar!(
+                r#"SELECT EXISTS(SELECT 1 FROM manifest WHERE digest = $1)"#,
                 mani_digest_str
             )
             .fetch_one(&mut *db.acquire().await?)
-            .await?
-                == 1;
-            if have_manifest {
-                return Ok(mani_digest);
+            .await?;
+            if has_manifest == 1 {
+                return Ok(mani_digest.to_string());
             }
             if let Some((cl, auth)) = &try_cl {
                 let ref_to_dl = image_ref.clone_with_digest(mani_digest.to_string());
@@ -181,24 +185,23 @@ impl SingleRegistryProxyConfig {
                 )
                 .await;
 
-                if let Err(e) = manifest_download {
-                    event!(Level::WARN, "Failed to download proxied image: {}", e)
-                } else {
-                    if let Some(tag) = image_ref.tag() {
-                        sqlx::query!(
-                            r#"
-                            INSERT INTO tag (repo, tag, manifest_digest)
-                            VALUES ($1, $2, $3)
-                            ON CONFLICT (repo, tag) DO UPDATE SET manifest_digest = $3
-                            "#,
-                            repo_name,
-                            tag,
-                            mani_digest_str
-                        )
-                        .execute(&mut *db.acquire().await?)
-                        .await?;
+                match manifest_download {
+                    Err(e) => tracing::warn!("Failed to download proxied image: {}", e),
+                    Ok(()) => {
+                        if let Some(tag) = image_ref.tag() {
+                            sqlx::query!(
+                                r#"INSERT INTO tag (repo, tag, manifest_digest)
+                                VALUES ($1, $2, $3)
+                                ON CONFLICT (repo, tag) DO UPDATE SET manifest_digest = $3"#,
+                                repo_name,
+                                tag,
+                                mani_digest_str
+                            )
+                            .execute(&mut *db.acquire().await?)
+                            .await?;
+                        }
+                        return Ok(mani_digest.to_string());
                     }
-                    return Ok(mani_digest);
                 }
             }
         }
@@ -208,7 +211,7 @@ impl SingleRegistryProxyConfig {
 }
 
 #[derive(thiserror::Error, Debug)]
-enum EcrPasswordError {
+pub enum EcrPasswordError {
     #[error("Could not parse region from ECR URL")]
     InvalidRegion,
     #[error("Could not decode ECR token: {0}")]
@@ -253,8 +256,7 @@ async fn get_aws_ecr_password_from_env(ecr_host: &str) -> Result<String, EcrPass
     Ok(String::from_utf8(auth_str)?)
 }
 
-/// `ref_` MUST reference a digest (*not* a tag)
-// #[async_recursion]
+/// `ref_` MUST reference a digest (_not_ a tag)
 async fn download_manifest_and_layers(
     cl: &oci_client::Client,
     auth: &RegistryAuth,
@@ -262,7 +264,7 @@ async fn download_manifest_and_layers(
     storage: &TrowStorageBackend,
     ref_: &Reference,
     local_repo_name: &str,
-) -> Result<()> {
+) -> Result<(), DownloadRemoteImageError> {
     async fn download_blob(
         cl: &oci_client::Client,
         db: SqlitePool,
@@ -270,8 +272,8 @@ async fn download_manifest_and_layers(
         ref_: &Reference,
         layer_digest: &str,
         local_repo_name: &str,
-    ) -> Result<()> {
-        event!(Level::TRACE, "Downloading blob {}", layer_digest);
+    ) -> Result<(), DownloadRemoteImageError> {
+        tracing::trace!("Downloading blob {}", layer_digest);
         let already_has_blob = sqlx::query_scalar!(
             "SELECT EXISTS(SELECT 1 FROM blob WHERE digest = $1);",
             layer_digest,
@@ -287,7 +289,7 @@ async fn download_manifest_and_layers(
                 .await?;
             let size = path.metadata().unwrap().len() as i64;
             sqlx::query!(
-                "INSERT INTO blob (digest, is_manifest, size) VALUES ($1, FALSE, $2) ON CONFLICT DO NOTHING;",
+                "INSERT INTO blob (digest, size) VALUES ($1, $2) ON CONFLICT DO NOTHING;",
                 layer_digest,
                 size
             )
@@ -301,14 +303,6 @@ async fn download_manifest_and_layers(
         )
         .execute(&mut *db.acquire().await?)
         .await?;
-        let parent_digest = ref_.digest().unwrap();
-        sqlx::query!(
-                "INSERT INTO blob_blob_association (parent_digest, child_digest) VALUES ($1, $2) ON CONFLICT DO NOTHING;",
-                parent_digest,
-                layer_digest
-            )
-            .execute(&mut *db.acquire().await?)
-            .await?;
 
         Ok(())
     }
@@ -320,7 +314,7 @@ async fn download_manifest_and_layers(
         oci_client::manifest::OCI_IMAGE_INDEX_MEDIA_TYPE,
     ];
 
-    event!(Level::DEBUG, "Downloading manifest + layers for {}", ref_);
+    tracing::debug!("Downloading manifest + layers for {}", ref_);
 
     let (raw_manifest, digest) = cl
         .pull_manifest_raw(ref_, auth, MIME_TYPES_DISTRIBUTION_MANIFEST)
@@ -329,58 +323,37 @@ async fn download_manifest_and_layers(
         oci_client::errors::OciDistributionError::ManifestParsingError(e.to_string())
     })?;
 
-    let manifest_size = raw_manifest.len() as i64;
-    let already_has_manifest = match sqlx::query!(
-        "INSERT INTO blob (digest, is_manifest, size) VALUES ($1, TRUE, $2);",
-        digest,
-        manifest_size
-    )
-    .execute(&mut *db.acquire().await?)
-    .await
-    {
-        Err(sqlx::Error::Database(e)) if e.is_unique_violation() => true,
-        Err(e) => return Err(e.into()),
-        Ok(_) => false,
-    };
-
-    sqlx::query!(
-        "INSERT INTO repo_blob_association (repo_name, blob_digest) VALUES ($1, $2) ON CONFLICT DO NOTHING;",
-        local_repo_name,
-        digest,
-    )
-    .execute(&mut *db.acquire().await?)
-    .await?;
-
-    if !already_has_manifest {
-        storage
-            .write_image_manifest(
-                Bytes::from(raw_manifest),
-                local_repo_name,
-                &Digest::try_from_raw(&digest).unwrap(),
-            )
-            .await?;
-    }
-
-    match manifest {
-        OCIManifest::List(m) => {
-            let images_to_dl = m
-                .manifests()
+    match &manifest {
+        OCIManifest::List(_) => {
+            let images_to_dl = manifest
+                .get_local_asset_digests()
                 .iter()
-                .map(|m| ref_.clone_with_digest(m.digest().as_ref().to_string()))
+                .map(|digest| ref_.clone_with_digest(digest.to_string()))
                 .collect::<Vec<_>>();
             let futures = images_to_dl.iter().map(|img| {
                 download_manifest_and_layers(cl, auth, db.clone(), storage, img, local_repo_name)
             });
             try_join_all(futures).await?;
         }
-        OCIManifest::V2(m) => {
-            let layer_digests = m.layers().iter().map(|layer| layer.digest().as_ref());
+        OCIManifest::V2(_) => {
+            let layer_digests = manifest.get_local_asset_digests();
             let futures = layer_digests
-                .clone()
+                .iter()
                 .map(|l| download_blob(cl, db.clone(), storage, ref_, l, local_repo_name));
             try_join_all(futures).await?;
         }
     }
+
+    sqlx::query!(
+        r"INSERT INTO manifest (digest, json, blob) VALUES ($1, jsonb($2), $2) ON CONFLICT DO NOTHING;
+        INSERT INTO repo_blob_association (repo_name, blob_digest) VALUES ($3, $4) ON CONFLICT DO NOTHING;",
+        digest,
+        raw_manifest,
+        local_repo_name,
+        digest
+    )
+    .execute(&mut *db.acquire().await?)
+    .await?;
 
     Ok(())
 }