+++ src/types.rs
@@ -1,7 +1,7 @@
 use std::ops::Deref;
 
-use futures::AsyncRead;
 use serde::{Deserialize, Serialize};
+use tokio::io::AsyncRead;
 
 use crate::registry::Digest;
 
@@ -47,7 +47,7 @@ pub struct DigestQuery {
 pub struct UploadInfo {
     uuid: String,
     repo_name: String,
-    range: (u32, u32),
+    range: (u64, u64),
 }
 
 pub struct BlobDeleted {}
@@ -55,7 +55,7 @@ pub struct BlobDeleted {}
 pub struct ManifestDeleted {}
 
 impl UploadInfo {
-    pub fn new(uuid: String, repo_name: String, range: (u32, u32)) -> Self {
+    pub fn new(uuid: String, repo_name: String, range: (u64, u64)) -> Self {
         Self {
             uuid,
             repo_name,
@@ -71,7 +71,7 @@ impl UploadInfo {
         &self.repo_name
     }
 
-    pub fn range(&self) -> (u32, u32) {
+    pub fn range(&self) -> (u64, u64) {
         self.range
     }
 }
@@ -81,11 +81,11 @@ pub struct AcceptedUpload {
     digest: Digest,
     repo_name: String,
     uuid: uuid::Uuid,
-    range: (u32, u32),
+    range: (u64, u64),
 }
 
 impl AcceptedUpload {
-    pub fn new(digest: Digest, repo_name: String, uuid: uuid::Uuid, range: (u32, u32)) -> Self {
+    pub fn new(digest: Digest, repo_name: String, uuid: uuid::Uuid, range: (u64, u64)) -> Self {
         Self {
             digest,
             repo_name,
@@ -102,7 +102,7 @@ impl AcceptedUpload {
         &self.repo_name
     }
 
-    pub fn range(&self) -> (u32, u32) {
+    pub fn range(&self) -> (u64, u64) {
         self.range
     }
 
@@ -123,16 +123,28 @@ pub struct VerifiedManifest {
     repo_name: String,
     digest: Digest,
     tag: String,
+    subject: Option<String>,
 }
 
 impl VerifiedManifest {
-    pub fn new(base_url: Option<String>, repo_name: String, digest: Digest, tag: String) -> Self {
+    pub fn new(
+        base_url: Option<String>,
+        repo_name: String,
+        digest: Digest,
+        tag: String,
+        subject: Option<String>,
+    ) -> Self {
         Self {
             base_url,
             repo_name,
             digest,
             tag,
+            subject,
         }
+    }
+
+    pub fn subject(&self) -> Option<&String> {
+        self.subject.as_ref()
     }
 
     pub fn digest(&self) -> &str {