+++ tests/proxy_registry.rs
@@ -21,35 +21,38 @@ mod proxy_registry {
     use crate::common::trow_router;
 
     async fn start_trow(data_dir: &Path) -> (Arc<TrowServerState>, Router) {
-        let config_file = RegistryProxiesConfig {
-            offline: false,
-            registries: vec![
-                SingleRegistryProxyConfig {
-                    alias: "docker".to_string(),
-                    host: "registry-1.docker.io".to_string(),
-                    username: None,
-                    password: None,
-                    ignore_repos: vec![],
-                },
-                SingleRegistryProxyConfig {
-                    alias: "nvcr".to_string(),
-                    host: "nvcr.io".to_string(),
-                    username: None,
-                    password: None,
-                    ignore_repos: vec![],
-                },
-                SingleRegistryProxyConfig {
-                    alias: "quay".to_string(),
-                    host: "quay.io".to_string(),
-                    username: None,
-                    password: None,
-                    ignore_repos: vec![],
-                },
-            ],
+        let config_file = trow::registry::ConfigFile {
+            registry_proxies: RegistryProxiesConfig {
+                offline: false,
+                registries: vec![
+                    SingleRegistryProxyConfig {
+                        alias: "docker".to_string(),
+                        host: "registry-1.docker.io".to_string(),
+                        username: None,
+                        password: None,
+                        ignore_repos: vec![],
+                    },
+                    SingleRegistryProxyConfig {
+                        alias: "nvcr".to_string(),
+                        host: "nvcr.io".to_string(),
+                        username: None,
+                        password: None,
+                        ignore_repos: vec![],
+                    },
+                    SingleRegistryProxyConfig {
+                        alias: "quay".to_string(),
+                        host: "quay.io".to_string(),
+                        username: None,
+                        password: None,
+                        ignore_repos: vec![],
+                    },
+                ],
+            },
+            ..Default::default()
         };
 
         trow_router(data_dir, |cfg| {
-            cfg.proxy_registry_config = Some(config_file);
+            cfg.config_file = Some(config_file);
         })
         .await
     }
@@ -170,12 +173,10 @@ mod proxy_registry {
         // Special case: docker/library
         // check that it works and that manifests are written in the correct location
         get_manifest(&trow, "f/docker/alpine", "3.13.4").await;
-        let digest = sqlx::query_scalar!("SELECT manifest_digest FROM tag WHERE repo = 'f/docker/library/alpine' AND tag = '3.13.4'")
+        sqlx::query_scalar!("SELECT manifest_digest FROM tag WHERE repo = 'f/docker/library/alpine' AND tag = '3.13.4'")
             .fetch_one(&mut *state.db.acquire().await.unwrap())
             .await
             .expect("Tag not found in database");
-        let file = data_dir.join(format!("./blobs/{digest}"));
-        assert!(file.exists());
     }
 
     #[tokio::test]